path_calculation_app/requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.0
h11==0.12.0
httptools==0.2.0
pydantic==1.8.2
python-decouple==3.4
python-dotenv==0.19.0
PyYAML==5.4.1
redis==3.5.3
rejson==0.5.4
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
uvloop==0.15.3
watchgod==0.7
websockets==9.1
